[prototype] Finished application concept
Added plain website that will serve as an interface for users to sent
search queries to the server and receive results. The site has already
linked JS script and CSS stylesheet files, but those two are also
minimal.
Connectivity to the C++ server is good enough to continue work with
the backend.